+++ src/resources/loader.h
@@ -25,8 +25,8 @@
 
 #include "../global.h"
 
-SDL_Surface* LoadSurface(const unsigned char* content, int size);
-SDL_Texture* LoadTexture(const unsigned char* content, int size);
-Mix_Chunk* LoadSound(const unsigned char* content, int size);
+SDL_Surface* LoadSurface(const unsigned char* content, size_t size);
+SDL_Texture* LoadTexture(const unsigned char* content, size_t size);
+Mix_Chunk* LoadSound(const unsigned char* content, size_t size);
 
 #endif
